Refactor shop management components and update asset references
- Updated ShopOrdersTab to use localized status labels for order statuses.
- Enhanced ShopProductsTab with a new product card component for better product display and management.
- Improved form handling and reset functionality in ShopProductsTab.
- Added new styles for shop product management layout and components.
- Updated asset references in index.html to use the latest JavaScript and CSS files.

+++ database/migrations/009_shop.sql
@@ -35,7 +35,7 @@ CREATE TABLE IF NOT EXISTS product_sizes (
 
 CREATE TABLE IF NOT EXISTS shop_orders (
   id BIGINT AUTO_INCREMENT PRIMARY KEY,
-  user_id BIGINT NULL,
+  user_id INT NULL,
   product_id BIGINT NOT NULL,
   assigned_shoe_id INT NULL,
   size VARCHAR(20) NULL,
